[svn-r333] Changes since 19980324 ----------------------
./MANIFEST
./src/H5HG.c		[new]
./src/H5HGprivate.h	[new]
./src/H5HGpublic.h	[new]
./src/Makefile.in
	Added global heaps, H5HG.  This compiles but is untested.

./src/H5D.c
./src/H5G.c
./src/H5Gnode.c
./src/H5Gstab.c
./src/H5HL.c		[renamed from H5H.c]
./src/H5HLprivate.h	[renamed from H5Hprivate.h]
./src/H5HLpublic.h	[renamed from H5Hpublic.h]
./src/H5Oefl.c
./src/Makefile.in
./src/hdf5.h
./test/theap.c
	Renamed local heaps from H5H to H5HL.

./src/H5AC.c
./src/H5ACprivate.h
./src/debug.c
	Added debugging support for global heap collections.

./html/H5.format.html
	Ripped out the old small object heap documentation and added
	documentation for the new global heap.

./html/H5.api.html
./src/H5P.c
./src/H5Ppublic.h
./src/H5F.c
./src/H5Fprivate.h
./src/H5private.h
	The H5Pget_version() function no longer returns a small object
	heap version number.  Instead, each collection of the global
	heap has its own version number instead of a single number in
	the file boot block.

./bin/release
	Omit files marked for removal by CVS even though the removal
	hasn't been committed yet.

./test/iopipe.c
	Removed warnings about casting unsigned int to double on Irix64.

./MANIFEST
	Removed ./test/testpar/phdf5sup.c

+++ test/iopipe.c
@@ -59,16 +59,22 @@ print_stats (const char *prefix,
 {
     double	u_time, s_time, e_time, bw;
 
-    u_time = (r_stop->ru_utime.tv_sec+r_stop->ru_utime.tv_usec/1000000.0) -
-	     (r_start->ru_utime.tv_sec+r_start->ru_utime.tv_usec/1000000.0);
+    u_time = ((double)(r_stop->ru_utime.tv_sec)+
+	      (double)(r_stop->ru_utime.tv_usec)/1000000.0) -
+	     ((double)(r_start->ru_utime.tv_sec)+
+	      (double)(r_start->ru_utime.tv_usec)/1000000.0);
 
-    s_time = (r_stop->ru_stime.tv_sec+r_stop->ru_stime.tv_usec/1000000.0) -
-	     (r_start->ru_stime.tv_sec+r_start->ru_stime.tv_usec/1000000.0);
+    s_time = ((double)(r_stop->ru_stime.tv_sec)+
+	      (double)(r_stop->ru_stime.tv_usec)/1000000.0) -
+	     ((double)(r_start->ru_stime.tv_sec)+
+	      (double)(r_start->ru_stime.tv_usec)/1000000.0);
 
-    e_time = (t_stop->tv_sec+t_stop->tv_usec/1000000.0) -
-	     (t_start->tv_sec+t_start->tv_usec/1000000.0);
+    e_time = ((double)(t_stop->tv_sec)+
+	      (double)(t_stop->tv_usec)/1000000.0) -
+	     ((double)(t_start->tv_sec)+
+	      (double)(t_start->tv_usec)/1000000.0);
 
-    bw = nbytes / e_time;
+    bw = (double)nbytes / e_time;
     
     printf (HEADING "%1.2fuser %1.2fsystem %1.2felapsed %1.2fMB/s\n", 
 	    prefix, u_time, s_time, e_time, bw/(1024*1024));
@@ -141,7 +147,8 @@ main (void)
     int			start[2];
     size_t		count[2];
     
-    printf ("I/O request size is %1.1fMB\n", (size[0]*size[1])/(1024.0*1024));
+    printf ("I/O request size is %1.1fMB\n",
+	    ((double)(size[0])*(double)(size[1]))/(1024.0*1024));
 
     
     /* Open the files */
